user-story-tests/22-inherit-all-outer-page-styles.html

<!DOCTYPE html>
<html lang="en">

<head>
    <meta charset="UTF-8">
    <meta name="viewport" content="width=device-width, initial-scale=1.0">
    <title>Inherit all outer page styles</title>
    <script src='./resources/testharness.js'></script>
    <script src='./resources/testharnessreport.js'></script>
</head>

<body>

    <style>
        button {
            border: thick solid black;
        }

        @layer buttons {
            button {
                border: thick dashed yellow;
            }
        }

        @import url(assets/resets.css) layer(resets);
    </style>
    <button>Button outside a shadow tree (thick solid black)</button>
    <button-group>
        <template shadowrootmode="open">
            <style>
                @layer inherit.layered.as.layered, inherit.unlayered.as.unlayered, layered, unlayered;
            </style>
            <button>
                Button inside a shadow tree (thick solid black) (from inherit.unlayered.as.unlayered)
            </button>
        </template>
    </button-group>

    <script type="module">
        import { adoptStylesFromLayerStatementRule } from "../AdoptStyles.js";

        window.onload = function () {

            const elName = "button-group"

            test(function () {

                const inheritResult = adoptStylesFromLayerStatementRule(document.querySelector(elName))
                assert_true(inheritResult, "inherit found, so returns true");
            }, "Inherit all outer page styles");

        }
    </script>

    <div id="log"></div>

</body>

</html>
